Upped the lower bound check for previous special attack to 90% to help limit notifications when using wells. Renamed package and primary plugin class and config interface to be the special attack not just special counter. Renamed notify method and previous percentage field to be more expressive.
